+++ app/Http/Controllers/ShippingAddressController.php
@@ -18,6 +18,14 @@ class ShippingAddressController extends Controller
     ];
 
     /**
+     * ShippingAddressController constructor.
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
      * Show the form for creating a new resource.
      *
      * @param \App\Company $company
@@ -51,12 +59,16 @@ class ShippingAddressController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param \App\Company $company
+     * @param  \App\Company  $company
      * @param  \App\ShippingAddress  $shippingAddress
+     *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Company $company, ShippingAddress $shippingAddress)
     {
+        $this->authorize('update', $shippingAddress);
+
         return view('shipping-address.edit')->with([
             'company' => $company,
             'shippingAddress' => $shippingAddress,
@@ -68,10 +80,14 @@ class ShippingAddressController extends Controller
      *
      * @param  \App\Company  $company
      * @param  \App\ShippingAddress  $shippingAddress
+     *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Company $company, ShippingAddress $shippingAddress)
     {
+        $this->authorize('update', $shippingAddress);
+
         $shippingAddress->update(request()->validate(self::RULES));
 
         if (request()->has('is_default')) {
@@ -84,12 +100,16 @@ class ShippingAddressController extends Controller
     /**
      * Show the form for deleting the specified resource.
      *
-     * @param \App\Company $company
-     * @param \App\ShippingAddress $shippingAddress
+     * @param  \App\Company  $company
+     * @param  \App\ShippingAddress  $shippingAddress
+     *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function delete(Company $company, ShippingAddress $shippingAddress)
     {
+        $this->authorize('delete', $shippingAddress);
+
         return view('shipping-address.delete')->with([
             'company' => $company,
             'shippingAddress' => $shippingAddress,
@@ -101,12 +121,15 @@ class ShippingAddressController extends Controller
      *
      * @param  \App\Company  $company
      * @param  \App\ShippingAddress  $shippingAddress
-     * @return \Illuminate\Http\Response
      *
+     * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @throws \Exception
      */
     public function destroy(Company $company, ShippingAddress $shippingAddress)
     {
+        $this->authorize('delete', $shippingAddress);
+
         $shippingAddress->delete();
 
         return redirect()->route('companies.show', ['company' => $company->id]);
@@ -117,10 +140,14 @@ class ShippingAddressController extends Controller
      *
      * @param  \App\Company  $company
      * @param  \App\ShippingAddress  $shippingAddress
+     *
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function default(Company $company, ShippingAddress $shippingAddress)
     {
+        $this->authorize('default', $shippingAddress);
+
         if (request()->has('is_default')) {
             $shippingAddress->default();
         }
